Move ReportingConfig to zigpy.quirks.v2
